*: split docker.go for future pkg creation
